fix(api): Include current release for all envs if none are filtered
Change the contract for GroupDetailsEndpoint.get such that
"currentRelease" is always included. It may still be null (with the same
meaning as before), but previously it would be omitted entirely unless
the request was filtered to one or more environments. Now, if there is
no filtering by environment, it will find a release across all
environments for the associated project.

The query that populates this field is a performance liability, and
making the query more often (that is, on all requests to the endpoint,
rather than the minority that filter by environment) will make the
performance cost more prevalent. We probably should not make this change
on production until the performance problem is solved, either by
optimizing the query or breaking it into its own endpoint.